Fixes Ticket #108. Also hides buildings.
Motivation behind hiding buildings: while it'd be nice as a layer toggle, showing buildings (based off OSM data) is kinda confusing.  The OSM data is very incomplete and we want people to draw buildings themselves on the map.  So the primary advantage to showing the buildings is to allow for easy tracing.  Eventually we'll have a layer toggle to show buildings (or better yet, aerial photography) as a layer for tracing purposes.