MergeModelForm -> MergeModelFormMixin.
This was a necessary change due to the limitations of subclassing ModelForms.  There's really no best practice for multiple inheritance with ModelForms, and because our ModelForm subclass (MergeModelForm) is so light it makes sense to adapt it to being a mixin.

See http://code.djangoproject.com/ticket/7018 for more information.  It didn't seem possible for me to use the olwidget ModelForm with MergeModelForm without some serious hackery.  So a mixin was in order.