Fix an issue w/ stylescombo & pages' native css.
When DEBUG=False, we use the combined ckeditor minified .js.  That minified .js has the stylescombo rolled into it.  But our inheritcss plugin, which loads the pages' css, should run before the stylescombo plugin creates the CSS path it uses.  In the case of the bunded stylescombo, this doesn't happen.

This is a workaround to that issue.  By making the plugin separate, we enforce the load order that's outlined in pages/widgets.py (the extra plugins function)