OpenLayers bump to 2.11. Permalink in map URL.
The map will now auto-update the URL as it's panned, zoomed, layer-switched, etc. OpenLayers 2.11 was necessary in order to get the anchor-in-URL behavior (rather than querystring).

On the initial map load there's no permalink -- it doesn't appear until the map is moved.  This is so that the URL is more general at first.