Pretty resource_uris. Allow lookup by any slugify-able slug.
In theory, this approach is bad.  We're really supposed to only have -one- URL for each resource.  But we're doing it this way in our apps, so it'd be better to keep this consistent and change the API and the apps at the same time.  The change would be to 301 Redirect to the page.pretty_slug whenever you visit anything that isn't page.pretty_slug.