Made tag resource lookup work correctly.
I created two new options for the SlugifyMixin -- lookup_function and to_url_function.  This is to allow custom slugify and URL encoding methods on resources.  We use a different slugify on tags and pages so we need some way of doing this.

Additionally, there is now a PageSlugifyMixin which uses the page slugify and name_to_url methods.  We'll use PageSlugifyMixin on anything that's got a page slug in the URL -- maps, pages, files (so far).