+ ModelHistoryResource. ServerStartupMiddleware.
Created a ModelHistoryResource to hold (for now) the history_user field.  In the process of trying to add this in, though, I repeatedly hit circular import errors.  These errors ultimately stemmed from the usage of models.py as the dumping-ground for post-server-startup register() calls -- e.g. for feeds, signals, and now the API.  Because so many different things depend on models, the easiest way to fix this was to find a way to run startup code /after/ the server has started up, but not during models.py import.  The best way to do this, until this officially lands as a post_startup signal in Django 1.5, is using this really hackish ServerStartupMiddleware thing.

So, now each app can have an optional startup.py, whose contents will be imported after the server has started up.