Fix disappearing map points on hover.
Fixes #460.

This isn't ideal, but it works.  Ideally we'd have a select even triggered by
hover, but I was having some trouble getting that working correctly.  The
select event seems to move the map around, causing more points to load, which
confuses the left-hand sidebar.  Should come back to this later and refactor.