Autocomplete pagename via haystack.
We now use haystack (and thus solr) to do pagename autocomplete. This will provide increased performance and better result matching. Before, we would only match on name__istartswith on pagename, but now we'll match substrings. I am sure there is room for improvement here, but this should be better than it was.

Also added a data migration that runs the update_index management command.