User pages. Mostly. Kind of.
For #656. There are a few UI and misc tweaks needed here, but this is
a good first step.

We need to unify user pages, as the underlying page content is currently
being stored in (kind of random) regions so that people can see user
page changes more easily.  This should be done after #42, for reasons
that are explained in #687