Multiple object types paginated in one view.
This is tricky.  We want to paginate through a sorted, merged
set of querysets.  We were able to do this before efficiently
only because we put a hard limit on the number of displayed
results (e.g. 500, or 7-days worth of changes).  Because we
now allow for infinite-loading recent changes, and because
we may need this again in the future to merge together
and paginate different types of objects, sorted by date, I
hacked this together.

It may need some tweaking, but I want to wait until we replicate
this behavior elsewhere before going much further.

This will speed up the RC loading speed on older regions (with
lots of pages over time) significantly.