Fix an edge case w/ reverting.
If a page had been deleted, then recreated, then a map added, then
renamed, then the original page reverted to the version before the map
was added --> error.  Unfortunately, I can't seem to reproduce this in a
test case, but can easily reproduce this with manual testing.  Will
investigate why the test case doesn't work later - I think it has to do
with transaction differences between the test and the actual view code,
somehow.

This was being caused by us not looking at the /closest/ version right
before the deletion for the related objects that we want to do a
cascade-revert over.